+++ HackU07-ios/Extensions/ExtensionUIView.swift
@@ -1,0 +1,33 @@
+//
+//  ExtensionUIView.swift
+//  HackU07-ios
+//
+//  Created by Etsushi Otani on 2022/06/21.
+//
+
+import Foundation
+import UIKit
+
+extension UIView {
+    func parentViewController() -> UIViewController? {
+        var parentResponder: UIResponder? = self
+        while true {
+            guard let nextResponder = parentResponder?.next else { return nil }
+            if let viewController = nextResponder as? UIViewController {
+                return viewController
+            }
+            parentResponder = nextResponder
+        }
+    }
+
+    func parentView<T: UIView>(type _: T.Type) -> T? {
+        var parentResponder: UIResponder? = self
+        while true {
+            guard let nextResponder = parentResponder?.next else { return nil }
+            if let view = nextResponder as? T {
+                return view
+            }
+            parentResponder = nextResponder
+        }
+    }
+}
