+++ HackU07-ios/Screen/News/NewsTableView.swift
@@ -67,27 +67,3 @@ class NewsTableView: UITableView, UITableViewDelegate, UITableViewDataSource {
             .store(in: &cancellables)
     }
 }
-
-extension UITableView {
-    func parentViewController() -> UIViewController? {
-        var parentResponder: UIResponder? = self
-        while true {
-            guard let nextResponder = parentResponder?.next else { return nil }
-            if let viewController = nextResponder as? UIViewController {
-                return viewController
-            }
-            parentResponder = nextResponder
-        }
-    }
-
-    func parentView<T: UIView>(type _: T.Type) -> T? {
-        var parentResponder: UIResponder? = self
-        while true {
-            guard let nextResponder = parentResponder?.next else { return nil }
-            if let view = nextResponder as? T {
-                return view
-            }
-            parentResponder = nextResponder
-        }
-    }
-}
